[Feat]: basic async syntax

+++ 0x01-python_async_function/0-basic_async_syntax.py
@@ -4,7 +4,7 @@ import random
 import asyncio
 
 
-async def wait_random(max_delay: int = 10) -> int:
+async def wait_random(max_delay: int = 10) -> float:
     """Wait for random integers between 0 and max_delay
 
     Args:
